+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `export(..)` as an experimental function, issue #104.
 * Fixed bug in formatting closure definitions, issue #103.
 * Enable `debug` logging for deprecated functions, issue #106.
+* Fixed bug in garbage collection, issue #107.
 
 # v0.9.18